// src/types.d.ts
export interface Settings {
	url?: string;
	apiKey: string;
}

// API types
/**
 * Model corresponding to https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/Comment
 */
export interface APIComment {
	comment_key: string;
	videoId: string;
	message: string;
}

/**
 * Model corresponding to https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/ChannelMin
 */
export interface APIChannelMin {
	id: string;
	name: string;
	englishName?: string;
	type: 'vtuber' | 'subber';
	photo?: string;
}

/**
 * Model corresponding to https://holodex.stoplight.io/docs/holodex/holodex_v2.yaml/components/schemas/Channel
 */
export interface APIChannel extends APIChannelMin {
	org?: string;
	subOrg?: string;
	banner?: string;
	twitter?: string;
	video_count?: string;
	subscriber_count?: string;
	view_count?: string;
	clip_count?: string;
	lang?: string;
	published_at: string;
	inactive: boolean;
	description: string;
}
// ...
export interface ListChannelParameters {
	lang?: 'en' | 'ja' | 'zh';
	limit?: number;
	offset?: number;
	order?: 'asc' | 'desc';
	org?: string;
	sort?: string;
	type?: 'subber' | 'vtuber';
}